refactor: move method
