+++ src/main/domain/item.ts
@@ -1,6 +1,7 @@
 import path from "path";
-import { readdirSync, statSync } from "fs";
+import { statSync } from "fs";
 import { ContentType } from "../../types";
+import { recursiveReaddir } from "../infrastructure/fileSystem";
 
 const extTypes = {
     image: ['.gif', '.jpg', '.jpeg', '.png', '.webp'],
@@ -27,18 +28,4 @@ export const specifyContentType = (location: string): ContentType => {
     if (extTypes.video.includes(ext)) return 'video';
 
     return 'other';
-}
-
-export const recursiveReaddir = (location: string): string[] => {
-    const fileOrDir = readdirSync(location);
-    const files: string[] = [];
-    fileOrDir.forEach(name => {
-        const fullpath = path.join(location, name);
-        if (statSync(fullpath).isDirectory()) {
-            files.push(...recursiveReaddir(fullpath));
-        } else {
-            files.push(fullpath);
-        }
-    });
-    return files;
 }
